Build Order Manager now issues build commands for buildings 4 seconds earlier than needed so the worker has to get into position. Why 4 seconds? Its usually a good amount of time, and the Build manager does not know the exact build site of the building, so a more accurate calculation based on travel distance and worker's max move speed cannot be done right now.
Also made a few changes to the construction manager so it can accept and process the build command, even though it doesn't have enough resources. So the construction manager now tries to get a probe into position as soon as possible, regardless of the amount of resources.

This update does not affect Terran addons or the production or morph managers.