TaskStream now can plan 4 tasks into the future. This allows the MacroSupplyManager to be implemented exactly based on the resource time line.
I also updated the UnitCompositionProducer so that it makes new task streams before the production buildings are complete. This enables the macro supply manager to consider the units that will be created from new barracks and factories.

I will need to add a special case for the first supply depot as it usually builds the first depot at around 7 or 8 supply to avoid getting supply blocked :P.