Broke just about everything in the MacroAIModule in a first step to move to a more generalized macro framework. Workers and TaskStreams with have a many-to-many relationship, so multiple workers will be able to work together to execute task streams, and workers will be able to handle multiple tasks streams. Code compiles, but I wouldn't try to run it yet. Most stuff is commented out for the moment to get things to compile.
Also added a new type of object called the WorkBench.

The number of WorkBenches a TaskStream has is the maximum number of workers that can work on the task stream. For example, if you have just one work bench, only 1 worker can execute tasks in the task stream. If you have 100 work benches, then up to 100 workers can execute tasks in the task stream. Only workbenches with workers can execute tasks. WorkBenches belong to TaskStreams.

2 examples:

If you have a task stream with 10 barracks and 2 work benches, then up to 2 SCVs will be used to make the 10 barracks. The basic worker finder should find a worker for each work bench once it is working again.

If you have a task stream with 100 marines and 100 work benches, and the basic worker finder is attached, then all available barracks should be used to train marines.

Once this is done it should make it relatively easy to implement the build order manager using the macro manager.